Clean up Bucket interface, move progress bars up to ConsoleTool.
The Bucket now reports upload/download progress to an object that
implements the ProgressListener interface.  The ConsoleTool makes
the progress bar object and passes it into the bucket.

There is now an UploadSource interface which is where the Bucket
gets data to upload, and a DownloadDestination where it puts
data downloaded from B2.

The code for downloading files still needs to be moved out of
ConsoleTool and into B2RawApi and Bucket.