Rethink ThreadPool class.
I think that we'll need to have tasks be able to add tasks
to the queue.  This means that using the shutdown tokens
in join() won't work, because other tasks might get added
behind them.

So the join() method now uses Queue.join().  And there is a
separate shut_down() method that stops the threads after
all of the tasks have been completed.