Redo b2http.py
(1) The http actions are now in a class called B2Http.  This will make it easier to mock/stub.

(2) The WebApiError exception is gone.  I decided it was an extra layer that wasn't needed.  Instead, the B2Http class returns the right subclass of B2Error.  This avoids having to make up HTTP status codes for things that aren't HTTP errors.