Large file uploads now upload parts in parallel.
The B2Api class holds a thread pool, which is used by the
Bucket to upload parts.

While looking for a Future object in Python, I found
concurrent.futures (and the futures package backport) that
provides a thread pool and future objects.  So I deleted
the ThreadPool class I wrote last week.