Add tqdm as a required package.
Progress bars were added before b2 was a python package,
so it couldn't have dependencies.  Now that there are
dependencies, there is no reason to avoid depending
on tqdm.

The latest tqdm package seemed to trigger more test
errors having to do with utf-8 encoding.  Fixed a few
issues in that area.