sync: handle out-of-order versions with --keepDays
The algorithm for deciding which versions to retain with
`--keepDays` requires that the versions be handled in
reverse-chronological order. However, this may not be the
case if a previous version was added with `--replaceNewer`,
which would trigger an assertion.

The goal of the original is to walk backwards and flip on a
"deleting" flag, but only _after_ we've walked past the most
recent version from before `keepDays` (which itself was the
state right _at_ `keepDays`, and should be retained). If our
times are out of order, though, we can't rely on a simple
flag. Instead, we have to record our current notion of this
"sentinel" version to keep, and be ready to replace it if we
find something newer.