Move filtering on sync to Folder.all_files method
With this implementation we avoid making a list of a whole
directory if we want to exclude some files/dirs.
API of the all_files method changed a bit but it was necessary.
Tests needed to be modified as well due to API changes.
All should work fine. All tests are passing.