Fix for #334 - Missing characters when syncing root
A couple people have taken stabs at fixing this one, but they
did not go far enough.  My original approach to this was misguided
because it confounded the path on the local file system and the
path in B2.  Building up both paths while descending through
the directory tree make the code much simpler.