Design change, more tests, add erm command
According to @mzukowski suggestions, the commands
has been completely reimplemented. Usage is different,
behavior is different also. There are 2 separate
commands now. rm is for glob-based removal and
erm is for regex-based one.
There is also a plan to improve current argument parsing
system to allow multiple arguments on the end of the line
but this will be another PR.