Fix build for 3.3 by not building for 3.3.
setuptools dropped support for 3.3 and refuses to
install anything in 3.3, so we're not able to run tests for it.

Also, stop claiming to support 3.3.  "If it's not tested, it doesn't work."

For some reason, setuptools does still support 2.6, which
hasn't had any releases since 2013.  I'm not taking that out
yet.