Fix #485
In `get_bucket_by_name`, do not try to list all of the buckets
because the application key may not allow that.

Also, remove the temporary workaround for the B2 service bug where
searching for buckets by name did not work.  The fix will be released
later today.

yapf for the test case.