Ensure tests in CI use the SDK from the master branch
It appears the intention in noxfile.py was to always run unit tests in
the CI with the latest SDK version, downloaded straight from git.

This didn't work, because right after installing the SDK from git, the
installation got overwritten with the `pip install -e .` command.

This also introduces changes which ensure that `liccheck` is run with
the SDK version that has been installed.