Moving changelog action to the last working version
Recent upgrade of dependencies of changelog changed the default behaviour
of unhandled promise rejection from just a warning to returning an error.
For a reason unknown, this happens only on master branch.

Issue was raised with the action creator, until it is resolved, an older
version of the action is used.