Fix for UnicodeEncodeError on Windows with Python 3.11
repr(args) still raised UnicodeEncodeError, this could be due to
removal of Py_UNICODE API. To combat this, we're printing pre-encoded
arguments – less readable, but working.

Note that this doesn't change the fact, that application is bundled
and signed using bash terminal, which should handle utf-8 quite well.
Issue presents itself only on the default powershell.